chore: remove constants covered in starknet api from papyrus
