+++ crates/starknet_api/src/transaction_hash.rs
@@ -29,6 +29,10 @@ use crate::transaction::{
 };
 use crate::{StarknetApiError, TransactionOptions};
 
+#[cfg(test)]
+#[path = "transaction_hash_test.rs"]
+mod transaction_hash_test;
+
 type ResourceName = [u8; 7];
 
 const DATA_AVAILABILITY_MODE_BITS: usize = 32;
